fix: modify the benchmark tool, move the original benchmark into legacy, because it will limit the threads to run
