+++ pytorch_extension/run.sh
@@ -1,5 +1,5 @@
 pip uninstall sparse-mm -y
 rm -rf build
 rm -rf dist
-rm -rf sparse-mm.egg-info
+rm -rf sparse_mm.egg-info
 python setup.py install
