+++ src/benchmark.py
@@ -1,31 +1,9 @@
-import torch
-import torch.utils.benchmark as benchmark
-from itertools import product
-from .extension import ExtensionHandler
 from .utils import generate_sparse_matrix, SparseMatrixTestConfiguration
+from src.benchmark_utils import benchmark, builtin_sparse_mm, builtin_sparse_mm_extension, openmp_sparse_mm, dense_mm, BenchmarkResult
 from typing import List
+from tqdm import tqdm
 
-def builtin_sparse_mm(sparse_matrix: torch.Tensor, sparse_matrix_1: torch.Tensor):
-    """Computes the product of two sparse matrices by using the builtin ``torch.sparse.mm`` function"""
-    return torch.sparse.mm(sparse_matrix, sparse_matrix_1)
-
-def builtin_sparse_mm_extension(sparse_matrix: torch.Tensor, sparse_matrix_1: torch.Tensor):
-    """Computes the product of two sparse matrices by using the extension version of ``torch.sparse.mm`` function"""
-    return ExtensionHandler.sparse_mm(sparse_matrix, sparse_matrix_1)
-
-def openmp_sparse_mm(sparse_matrix: torch.Tensor, sparse_matrix_1: torch.Tensor):
-    """Computes the product of two sparse matrices by using the openmp version of ``torch.sparse.mm`` function"""
-    return ExtensionHandler.openmp_sparse_mm(sparse_matrix, sparse_matrix_1)
-
-def dense_mm(sparse_matrix: torch.Tensor, sparse_matrix_1: torch.Tensor):
-    """Computes the product of two dense matrices by using the builtin ``torch.mm`` function"""
-    if sparse_matrix.is_sparse:
-        sparse_matrix = sparse_matrix.to_dense()
-    if sparse_matrix_1.is_sparse:
-        sparse_matrix_1 = sparse_matrix_1.to_dense()
-    return torch.mm(sparse_matrix, sparse_matrix_1)
-
-def benchmark_functions_single_test(config: SparseMatrixTestConfiguration, num_runs):
+def _benchmark(config: SparseMatrixTestConfiguration, num_runs: int):
     """
     Benchmark the builtin ``torch.sparse.mm`` function and the extension version of ``torch.sparse.mm`` function.
     Compare it on specific sizes of sparse matrices and specific densities.
@@ -33,101 +11,42 @@ def benchmark_functions_single_test(config: SparseMatrixTestConfiguration, num_r
     sparse_matrix = generate_sparse_matrix(config.A_row, config.A_col, density=config.A_density)
     sparse_matrix_1 = generate_sparse_matrix(config.B_row, config.B_col, density=config.B_density)
 
-    t0 = benchmark.Timer(
-        stmt='builtin_sparse_mm(sparse_matrix, sparse_matrix_1)',
-        setup='from src.benchmark import builtin_sparse_mm',
-        globals={'sparse_matrix': sparse_matrix, 'sparse_matrix_1': sparse_matrix_1})
+    t0 = 0.0
+    t1 = 0.0
+    t2 = 0.0
+    t3 = 0.0
 
-    t1 = benchmark.Timer(
-        stmt='builtin_sparse_mm_extension(sparse_matrix, sparse_matrix_1)',
-        setup='from src.benchmark import builtin_sparse_mm_extension',
-        globals={'sparse_matrix': sparse_matrix, 'sparse_matrix_1': sparse_matrix_1})
-    
-    t2 = benchmark.Timer(
-        stmt='openmp_sparse_mm(sparse_matrix, sparse_matrix_1)',
-        setup='from src.benchmark import openmp_sparse_mm',
-        globals={'sparse_matrix': sparse_matrix, 'sparse_matrix_1': sparse_matrix_1})
+    for i in range(num_runs):
+        t0 += benchmark(builtin_sparse_mm, sparse_matrix, sparse_matrix_1)
+        t1 += benchmark(builtin_sparse_mm_extension, sparse_matrix, sparse_matrix_1)
+        t2 += benchmark(openmp_sparse_mm, sparse_matrix, sparse_matrix_1)
+        t3 += benchmark(dense_mm, sparse_matrix.to_dense(), sparse_matrix_1.to_dense())
 
-    t3 = benchmark.Timer(
-        stmt='dense_mm(sparse_matrix.to_dense(), sparse_matrix_1.to_dense())',
-        setup='from src.benchmark import dense_mm',
-        globals={'sparse_matrix': sparse_matrix, 'sparse_matrix_1': sparse_matrix_1})
+    return BenchmarkResult(t0 / num_runs, t1 / num_runs, t2 / num_runs, t3 / num_runs, config.A_col, config.A_row, config.A_density, config.B_col, config.B_row, config.B_density)
 
-    print(f'builtin_sparse_mm(sparse_matrix, sparse_matrix_1):  {t0.timeit(num_runs)})')
-    print(f'builtin_sparse_mm_extension(sparse_matrix, sparse_matrix_1):      {t1.timeit(num_runs)})')
-    print(f'openmp_sparse_mm(sparse_matrix, sparse_matrix_1):      {t2.timeit(num_runs)})')
-    print(f'dense_mm(sparse_matrix.to_dense(), sparse_matrix_1.to_dense()):      {t2.timeit(num_runs)})')
-    print(f'Finsish Benchmarking with A_row={config.A_row}, A_col={config.A_col}, A_density={config.A_density}, B_row={config.B_row}, B_col={config.B_col}, B_density={config.B_density}, num_runs={num_runs}')
-    print("-" * 50 + "\n")
+def show_benchmark_results(results: List[BenchmarkResult]) -> None:
+    """
+    Print the benchmark results.
+    """
+    for result in results:
+        print(result)
 
-# WARNING: This function may cause CUDA kernel timeout errors if the benchmark function call time exceeds a certain threshold.
-# SOLUTION: Limit the number of configurations in the list to prevent this issue.
-def benchmark_functions_multiple_test(config_list: List[SparseMatrixTestConfiguration], results: List):
+def benchmark_by_config_list(config_list: List[SparseMatrixTestConfiguration], num_runs: int) -> List[BenchmarkResult]:
     """
     Benchmark the builtin ``torch.sparse.mm`` function and the extension version of ``torch.sparse.mm`` function.
     Compare it on different sizes of sparse matrices and different densities.
     """
-
-    for i, config in enumerate(config_list):
-        label = 'Benchmarking SPMM strategies'
-        sub_label = f'[{config.A_row:5}, {config.A_col:5}, density={config.A_density:5.2f}], [{config.B_row:5}, {config.B_col:5}, density={config.B_density:5.2f}]'
-        sparse_matrix = generate_sparse_matrix(config.A_row, config.A_col, density=config.A_density)
-        sparse_matrix_1 = generate_sparse_matrix(config.B_row, config.B_col, density=config.B_density)
-        # print(f'Builtin SPMM: {i*4 + 0}')
-        results.append(benchmark.Timer(
-            stmt='builtin_sparse_mm(sparse_matrix, sparse_matrix_1)',
-            setup='from src.benchmark import builtin_sparse_mm',
-            globals={'sparse_matrix': sparse_matrix, 'sparse_matrix_1': sparse_matrix_1},
-            label=label,
-            sub_label=sub_label,
-            description='builtin_sparse_mm').timeit(1))
-        
-        # print(f'Extension SPMM: {i*4 + 1}')
-        results.append(benchmark.Timer(
-            stmt='builtin_sparse_mm_extension(sparse_matrix, sparse_matrix_1)',
-            setup='from src.benchmark import builtin_sparse_mm_extension',
-            globals={'sparse_matrix': sparse_matrix, 'sparse_matrix_1': sparse_matrix_1},
-            label=label,
-            sub_label=sub_label,
-            description='builtin_sparse_mm_extension').timeit(1))
-        
-        # print(f'Parallel SPMM: {i*4 + 2}')
-        results.append(benchmark.Timer(
-            stmt='openmp_sparse_mm(sparse_matrix, sparse_matrix_1)',
-            setup='from src.benchmark import openmp_sparse_mm',
-            globals={'sparse_matrix': sparse_matrix, 'sparse_matrix_1': sparse_matrix_1},
-            label=label,
-            sub_label=sub_label,
-            description='openmp_sparse_mm').timeit(1))
-
-        # print(f'Dense MM: {i*4 + 2}')
-        results.append(benchmark.Timer(
-            stmt='dense_mm(sparse_matrix.to_dense(), sparse_matrix_1.to_dense())',
-            setup='from src.benchmark import dense_mm',
-            globals={'sparse_matrix': sparse_matrix, 'sparse_matrix_1': sparse_matrix_1},
-            label=label,
-            sub_label=sub_label,
-            description='dense_mm').timeit(1))
-        
-def show_benchmark_results(results):
-    """
-    Show the benchmark results.
-    """
-    compare = benchmark.Compare(results)
-    compare.print()
-
-def generate_and_benchmark_configurations(size_start, size_end, size_step, density_start, density_end, density_step):
-    """
-    Generate a list of test configurations and benchmark them.
-    """
-    size_range = range(size_start, size_end, size_step)
-    density_range = range(density_start, density_end, density_step)
     results = []
-    from tqdm import tqdm
-    for size in tqdm(size_range, desc="Size"):
-        configurations = []
-        for density in [i/10 for i in density_range]:
-            for density_2 in [i/10 for i in density_range]:
-                configurations.append(SparseMatrixTestConfiguration(size, size, density, size, size, density_2))
-        benchmark_functions_multiple_test(configurations, results)
+    for config in tqdm(config_list):
+        results.append(_benchmark(config, num_runs))
     return results
+
+def generate_benchmark_configurations(size_start: int, size_end: int, size_step: int, density_start: float, density_end: float, density_step: float) -> List[SparseMatrixTestConfiguration]:
+    """
+    Generate a list of benchmark configurations.
+    """
+    config_list = []
+    for size in range(size_start, size_end, size_step):
+        for density in range(density_start, density_end, density_step):
+            config_list.append(SparseMatrixTestConfiguration(size, size, density / 10, size, size, density / 10))
+    return config_list
